Move Verifier under File namespace
Now that we have Acl and List under the File namespace, it makes sense
for Verifier to live there as well.
