+++ test/gcloud/storage/test_verifier.rb
@@ -14,7 +14,7 @@
 
 require "helper"
 
-describe Gcloud::Storage::Verifier, :mock_storage do
+describe Gcloud::Storage::File::Verifier, :mock_storage do
   let(:file_contents) { "The quick brown fox jumps over the lazy dog." }
   let(:md5_digest) { "1B2M2Y8AsgTpgAmY7PhCfg==" }
   let(:crc32c_digest) { "AAAAAA==" }
@@ -26,21 +26,21 @@ describe Gcloud::Storage::Verifier, :mock_storage do
   it "verifies md5 digest" do
     Tempfile.open "gcloud-ruby" do |tmpfile|
       tmpfile.write file_contents
-      assert Gcloud::Storage::Verifier.verify_md5(file, tmpfile)
+      assert Gcloud::Storage::File::Verifier.verify_md5(file, tmpfile)
     end
   end
 
   it "verifies crc32c digest" do
     Tempfile.open "gcloud-ruby" do |tmpfile|
       tmpfile.write file_contents
-      assert Gcloud::Storage::Verifier.verify_crc32c(file, tmpfile)
+      assert Gcloud::Storage::File::Verifier.verify_crc32c(file, tmpfile)
     end
   end
 
   it "calculates md5 digest" do
     Tempfile.open "gcloud-ruby" do |tmpfile|
       tmpfile.write file_contents
-      digest = Gcloud::Storage::Verifier.md5_for tmpfile
+      digest = Gcloud::Storage::File::Verifier.md5_for tmpfile
       digest.must_equal md5_digest
     end
   end
@@ -48,7 +48,7 @@ describe Gcloud::Storage::Verifier, :mock_storage do
   it "calculates crc32c digest" do
     Tempfile.open "gcloud-ruby" do |tmpfile|
       tmpfile.write file_contents
-      digest = Gcloud::Storage::Verifier.crc32c_for tmpfile
+      digest = Gcloud::Storage::File::Verifier.crc32c_for tmpfile
       digest.must_equal crc32c_digest
     end
   end
